requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
groq
schedule
python-dotenv
langchain 
langchain-groq
langchain-community
langgraph
twilio
flask
fastapi
uvicorn
python-multipart
gspread 
oauth2client
google-search-results
flask-cors
markdown
